Adding my little files
May or may not refine them.
Alarm: Put in a time and number of times to repeat and it sets an alarm. 
Crawler: Just takes the headlines from MSNBC and returns them - I don't care about it in the slightest but I can see how I could easily extend it
URL/Update_Url: Opens a webpage when I restart my computer and allows me to update the url.txt file that I store it in. I made this to remind me to do certain projects, etc, after sleeping.